+++ multitenancy-demo-boot-parent/multitenancy-demo-webapp/src/main/java/com/metafour/multitenancy/bean/EmployeeRepository.java
@@ -1,5 +1,8 @@
 package com.metafour.multitenancy.bean;
 
+import java.util.List;
+
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.repository.CrudRepository;
 
 /**
@@ -11,4 +14,6 @@ import org.springframework.data.repository.CrudRepository;
  */
 public interface EmployeeRepository extends CrudRepository<Employee, Long> {
 
+	@Cacheable("cachedFirstName")
+	List<Employee> findByFirstName(String firstname);
 }